leftovers, NFC

+++ src/renderer/RenderDevice.cpp
@@ -507,8 +507,8 @@ void RenderDevice::RenderThread(RenderDevice* rd, const bgfx::Init& initReq)
             span* tagSpan = new span(series, 1, _T("WaitSync"));
             #endif
             U64 now = usec();
-            const int refreshLength = static_cast<int>(1000000.f / rd->m_outputWnd[0]->GetRefreshRate());
-            const int minimumFrameLength = static_cast<int>(1000000.f / g_pplayer->m_maxFramerate);
+            const int refreshLength = static_cast<int>(1000000. / (double)rd->m_outputWnd[0]->GetRefreshRate());
+            const int minimumFrameLength = static_cast<int>(1000000. / (double)g_pplayer->m_maxFramerate);
             const int maximumFrameLength = 5 * refreshLength;
             const int targetFrameLength = clamp(refreshLength - 2000, min(minimumFrameLength, maximumFrameLength), maximumFrameLength);
             while (now - rd->m_lastPresentFrameTick < targetFrameLength)
